requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2

# Google Cloud
google-cloud-storage==3.1.0
google-cloud-pubsub==2.29.0

# dbt
dbt-core==1.9.4
dbt-postgres==1.9.1

# LangChain + agents
langchain==0.3.21
langchain-openai
langchain-community==0.3.20
langchain-core==0.3.49
PyGithub==2.9.0

# Config & utilities
pydantic-settings==2.8.1
python-dotenv==1.1.0
httpx==0.28.1
